Ores Meta texture overlays.
These are pre-alpha and subject to changes.
Due to texture overlay alpha-channel restriction to 1bit on blocks,  and single tint hardcoded into mod, it is not possible to obtain same consistant Soartex styled ores.
Considering these limitations and mod specific ores designs and implementation. It may be an alternate good idea to draw something entirely different.