Tinkers Construct - Closes #22
- Fixed: The Text bug with the scroll over cutting the title causing it
to become un-readable.
- Fixed: The input slots using the wrong texture, it has to be either
this border or the alt not the light one. As the patterns become hard to
see or you have an overlapping bug when you hover over the slot seen
[here](https://dl.dropboxusercontent.com/u/17094065/ShareX/2014/11/2014-11-22_1920-1080-22-04-31-15.jpg)
- Added: Description GUI hard to work out this one but it looks awesome
and renders correctly in game.
- Tweaked: Added a background on the patternchestmini (For when attached
to the tool parts table) so the title is easier to see.
- Notes: Some other GUI's across many mods could have these small tweaks
to make the titles easier to see, as this is a common bug I've seen when
playing with a large range of mods.