Reversion to Mod Style on BOP Magic Tree
The Magic Tree in Soartex looks nothing like the one in the mod, and the
mod appearance is both attractive and not covered elsewhere in the mod.
I have reverted it to the blue/cyan color scheme with blue wood, which I
think looks very nice. There is already green wood in the mod,
specifically the willow.