Adds Woods and Leaves for untextured trees
Brings us to near completion for all woods, leaves, and saplings in BoP.
A palm tree sapling texture is included in Alts, as I apparently cannot
draw a palm tree that small with any quality. Does not include Etherial
items, might add those in a later commit, but wanted to cover all trees
that currently grow in the world because of this mod.